chore: refactor text field height and add disable submit checks
