+++ src/client/user/constants.ts
@@ -5,4 +5,4 @@ export const initialSortConfig = {
   sortDirection: SortDirection.Descending,
 }
 
-export default initialSortConfig
+export const TEXT_FIELD_HEIGHT = 44
